# docs/swagger.yaml
definitions:
  model.AdminUser:
    properties:
      avatar:
        description: 头像
        type: string
      created_at:
        type: integer
      id:
        description: ID
        type: integer
      is_open:
        description: 是否启用 1是 2否
        type: integer
      name:
        description: 名称
        type: string
      remember_token:
        description: 免登录token
        type: string
      user_name:
        description: 用户名
        type: string
    type: object
...
  response.Code:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
info:
  contact:
    email: xiangfudada@163.com
    name: Xfdada
  title: ICQQG接囗文档
  version: "1.0"
paths:
...
  /api/web/user:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":model.AdminUser}'
          schema:
            $ref: '#/definitions/model.AdminUser'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取自身信息
      tags:
      - 后端管理员模块
    post:
      parameters:
      - description: ' 用户名'
        in: formData
        name: username
        required: true
        type: string
      - description: ' 密码'
        in: formData
        name: password
        required: true
        type: string
      - description: ' 姓名'
        in: formData
        name: name
        required: true
        type: string
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"新增成功"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 新增管理员
      tags:
      - 后端管理员模块
  /api/web/user/{username}:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 账号名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.AdminUser}'
          schema:
            items:
              $ref: '#/definitions/model.AdminUser'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 删除管理员
      tags:
      - 后端管理员模块
...
  /api/web/userIsOpen:
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 账号名
        in: formData
        name: username
        required: true
        type: string
      - description: 是否启用
        in: formData
        name: is_open
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.AdminUser}'
          schema:
            items:
              $ref: '#/definitions/model.AdminUser'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 管理员是否启用
      tags:
      - 后端管理员模块
  /api/web/userList:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.AdminUser}'
          schema:
            items:
              $ref: '#/definitions/model.AdminUser'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取管理员列表
      tags:
      - 后端管理员模块
swagger: "2.0"